+++ modules/DHTsensors/DHTsensors.h
@@ -4,7 +4,7 @@
 #include <Arduino.h> // Include the Arduino core header file
 #include <DHT.h>
 
-#define DHTPIN 2            // what pin we're connected to
+#define DHTPIN 0            // what pin we're connected to
 //#define DHTTYPE DHT11     // DHT 11
 #define DHTTYPE DHT22       // DHT 22 (AM2302), AM2321
 //#define DHTTYPE DHT21     // DHT 21 (AM2301)