+++ modules/WS2812B_Matrix/WS2812B_Matrix.h
@@ -25,6 +25,7 @@ extern byte displayON;
 // Params for width and height
 extern const uint8_t kMatrixWidth;
 extern const uint8_t kMatrixHeight;
+extern byte kMatrixOrientation;
 
 // Param for different pixel layouts
 extern const bool kMatrixSerpentineLayout;
@@ -95,14 +96,12 @@ extern const uint8_t FontWidth;
 
 
 extern void updateDisplay();
-extern void drawLetter( int posx, int posy, char letter, CRGB color );
-extern void drawLetterBuf( CRGB *buffer, int posx, int posy, char letter, CRGB color );
+extern void drawLetter(int posx, int posy, char letter, CRGB color, int orientation, CRGB *buffer = nullptr);
 extern void drawTempHum( int x, int y, CRGB colorText, bool isTemperature );
 extern void drawTime( int x, int y, CRGB colorTime, bool colon, bool seconds );
 extern void drawDate( CRGB *buffer, int x, int y, CRGB colorDate );
-extern void displayMessage( CRGB colorScroll, const char *message, int numSpaces );
+extern void displayMessage(CRGB *buffer, CRGB colorScroll, const char *message, int numSpaces, int orientation);
 extern void renderDisplayMessage( unsigned long currentMillis );
-
 extern void drawWIN( const byte coords[][2], int size, CRGB *buffer, CRGB color );
 extern void drawText( CRGB *buffer );
 extern void renderDisplayWin( unsigned long currentMillis );